Factor out progress-step-route.tsx.
