+++ frontend/lib/progress-redirection.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { AllSessionInfo } from "./queries/AllSessionInfo";
-import { ProgressStepRoute } from "./progress-bar";
+import { ProgressStepRoute } from "./progress-step-route";
 import { withAppContext, AppContextType } from "./app-context";
 import { Redirect } from "react-router";
